requirements.txt
Django==5.0.8
djangorestframework==3.15.2
pillow==10.4.0
